Moving code around so it lines up with the ref assemblies. Getting the full build working again.
