+++ src/libraries/System.Utf8String.Experimental/ref/System.Utf8String.Experimental.Forwards.cs
@@ -1,0 +1,7 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Index))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Range))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.Rune))]
