+++ src/libraries/System.Utf8String.Experimental/ref/System.Utf8String.Experimental.cs
@@ -54,9 +54,6 @@ namespace System
         public static System.ReadOnlySpan<byte> AsBytes(this System.Utf8String? text) { throw null; }
         public static System.ReadOnlySpan<byte> AsBytes(this System.Utf8String? text, int start) { throw null; }
         public static System.ReadOnlySpan<byte> AsBytes(this System.Utf8String? text, int start, int length) { throw null; }
-        public static System.ReadOnlyMemory<System.Char8> AsMemory(this System.Utf8String? text) { throw null; }
-        public static System.ReadOnlyMemory<System.Char8> AsMemory(this System.Utf8String? text, int start) { throw null; }
-        public static System.ReadOnlyMemory<System.Char8> AsMemory(this System.Utf8String? text, int start, int length) { throw null; }
         public static System.ReadOnlyMemory<byte> AsMemoryBytes(this System.Utf8String? text) { throw null; }
         public static System.ReadOnlyMemory<byte> AsMemoryBytes(this System.Utf8String? text, int start) { throw null; }
         public static System.ReadOnlyMemory<byte> AsMemoryBytes(this System.Utf8String? text, System.Index startIndex) { throw null; }
