+++ src/libraries/System.Utf8String.Experimental/ref/System.Utf8String.Experimental.netcoreapp.cs
@@ -7,11 +7,6 @@
 
 namespace System
 {
-    public static partial class Utf8Extensions
-    {
-        public static System.ReadOnlyMemory<System.Char8> AsMemory(this System.Utf8String? text, System.Index startIndex) { throw null; }
-        public static System.ReadOnlyMemory<System.Char8> AsMemory(this System.Utf8String? text, System.Range range) { throw null; }
-    }
     public sealed partial class Utf8String : System.IComparable<System.Utf8String?>, System.IEquatable<System.Utf8String?>
     {
         public static System.Utf8String Create<TState>(int length, TState state, System.Buffers.SpanAction<byte, TState> action) { throw null; }
