+++ src/libraries/System.Utf8String.Experimental/src/System/ThrowHelper.cs
@@ -74,6 +74,7 @@ namespace System
         input,
         length,
         start,
+        text,
         value,
     }
 }
